Add XMLGenerator.go
An XMLGenerator will generate valid XML based off the input file and pass
it to mutators and harnesses. Currently XMLGenerator.go reads in xml input
from a file and parses it into a tree structure. The tree structure is
contained within an mXMLHolder struct which additionally holds a description
containg the changes the generator made to the original input to create the resulting xml that will be sent to the mutators and harnesses. Having an
XMLTree structure provides an easy way to add additional XML elements to
the original input.