Add spamElementBreadthWise generator function, remove BFSUtil function
This generator function produces valid XML based on XML from an input file.
The output from this function can be passed on to a mutator so further
mutations can be performed on it. xmltree provides a method Flatten() that takes in an Element tree and returns a slice of Elements. This does the same
this as BFSUtil. getElemPool uses Flatten() rather than BFSUtil.