Add permutator.go add permutator interface.
Require an interface so that main can call the one method to permutate
input despite there being different ways we want to permutate depending on
the format the input conforms to be it XML, JSON, CSV or plaintext.

The createPermutator method will return a struct that conforms to the
permutator interface.