Add chan to mutator in CSVPermutator.
The permutateInput method now passes permutated input to a channel which
the mutator grabs from and performs further mutations on.
The permutator interface has been modified to reflect this change and now
requires two channels in the permutateInput method argument. One channel
to send permutations to harnesses another channel to send permutations
to mutators.