Add a spam method to the JSONPermutor
Spam method will randomly select some child elements and call their spam
method. Originally planned on recursivly calling spam() method down to
an int or string, but this was too slow (or there is a bug that I
haven't found yet). The spam method does not deep copy objects, which
causes some issues, but can hopefully be ignored for the moment