Add permCSV - a csv permutator
A permCSV is a csv permutator consisting of two outbound channels and
a current permutation which is a parsed CSV file.

With this method, permutateInput can add as many different variations of
permutation functions as it wants. It simply sets currPerm to a newly
parsed file, passes currPerm to a function, then sends the currPerm
across the two outbound channels.

For performance, it might be worth looking into having the original
parsed file stored in the permutator so that the currPerm can be
created by deep copying the original parsed csv rather than opening
and reading from a file each time.