Add fileBytes field to mutator struct
permutators send input to a channel which a mutator receives from.
If a permtator has not been implemented yet for a file format, the mutator
will block forever waiting to receive data from the channel. With the addition of the fileBytes field to the mutator, the mutator can create a new TestCase using the fileBytes stored in its struct if it does not receive anything from the channel.