Change harness strategy to memory snapshotting
This will move the harnessing strategy away from successive fork+exec
cycles into a memory snapshot/restore system, where the harnessed
process's state is saved and subsequently restored after running. This
results in a large performance boost by avoiding the large overhead of
the usual process startup sequence.
Additionally, checks are put in place for the architecture of the target
binary, which is required for the various snapshotting/tracing
strategies to work. Support is added for x86 32/64bit binaries only.