Replace mutateStrings with insertString
mutateStrings was inflexible. It used a regex to find anything within double
quotes and replace whatever was in the double quotes with an interesting
string. This made it unlikely to perform any mutations on file formats like csv
or plaintext. Changing the regex to be more inclusive and passing that
to mutateObj was also impractical. The number candidates that matched a
regex such as [a-zA-Z]+ was large which meant mutateObj would mutate a large
number of these candidates, replacing each of them with a randomly chosen string from interestingString.
This was not ideal for file formats such as xml which require matching start
and end tags.
insertString simply selects an arbritrary location within ts.input and adds
a single string returned from interestingString. This is faster than using
mutateObj since it removes the need for regex. It reduces the chance of the
fuzzer producing invalid xml since only one string is inserted. Furthermore it will work with files in csv and plaintext format, testing them for vulns such
as format strings and buffer overflows.